+++ internal/database/metadata/postgres/tx.go
@@ -6,7 +6,7 @@ import (
 	"github.com/oom-ai/oomstore/internal/database/metadata"
 )
 
-func (tx *Tx) CreateEntity(ctx context.Context, opt metadata.CreateEntityOpt) (int16, error) {
+func (tx *Tx) CreateEntity(ctx context.Context, opt metadata.CreateEntityOpt) (int, error) {
 	return createEntity(ctx, tx, opt)
 }
 
@@ -14,7 +14,7 @@ func (tx *Tx) UpdateEntity(ctx context.Context, opt metadata.UpdateEntityOpt) er
 	return updateEntity(ctx, tx, opt)
 }
 
-func (tx *Tx) CreateFeatureGroup(ctx context.Context, opt metadata.CreateFeatureGroupOpt) (int16, error) {
+func (tx *Tx) CreateFeatureGroup(ctx context.Context, opt metadata.CreateFeatureGroupOpt) (int, error) {
 	return createFeatureGroup(ctx, tx, opt)
 }
 
@@ -22,7 +22,7 @@ func (tx *Tx) UpdateFeatureGroup(ctx context.Context, opt metadata.UpdateFeature
 	return updateFeatureGroup(ctx, tx, opt)
 }
 
-func (tx *Tx) CreateFeature(ctx context.Context, opt metadata.CreateFeatureOpt) (int16, error) {
+func (tx *Tx) CreateFeature(ctx context.Context, opt metadata.CreateFeatureOpt) (int, error) {
 	return createFeature(ctx, tx, opt)
 }
 
@@ -30,7 +30,7 @@ func (tx *Tx) UpdateFeature(ctx context.Context, opt metadata.UpdateFeatureOpt) 
 	return updateFeature(ctx, tx, opt)
 }
 
-func (tx *Tx) CreateRevision(ctx context.Context, opt metadata.CreateRevisionOpt) (int32, string, error) {
+func (tx *Tx) CreateRevision(ctx context.Context, opt metadata.CreateRevisionOpt) (int, string, error) {
 	return createRevision(ctx, tx, opt)
 }
