+++ pkg/oomstore/feature_group.go
@@ -7,14 +7,14 @@ import (
 	"github.com/oom-ai/oomstore/pkg/oomstore/types"
 )
 
-func (s *OomStore) CreateFeatureGroup(ctx context.Context, opt metadata.CreateFeatureGroupOpt) (int16, error) {
+func (s *OomStore) CreateFeatureGroup(ctx context.Context, opt metadata.CreateFeatureGroupOpt) (int, error) {
 	// Via the oomstore API, we can only create a batch feature group
 	// So we hardcode the category to be batch
 	opt.Category = types.BatchFeatureCategory
 	return s.metadata.CreateFeatureGroup(ctx, opt)
 }
 
-func (s *OomStore) GetFeatureGroup(ctx context.Context, id int16) (*types.FeatureGroup, error) {
+func (s *OomStore) GetFeatureGroup(ctx context.Context, id int) (*types.FeatureGroup, error) {
 	return s.metadata.GetFeatureGroup(ctx, id)
 }
 
@@ -22,7 +22,7 @@ func (s *OomStore) GetFeatureGroupByName(ctx context.Context, name string) (*typ
 	return s.metadata.GetFeatureGroupByName(ctx, name)
 }
 
-func (s *OomStore) ListFeatureGroup(ctx context.Context, entityID *int16) types.FeatureGroupList {
+func (s *OomStore) ListFeatureGroup(ctx context.Context, entityID *int) types.FeatureGroupList {
 	return s.metadata.ListFeatureGroup(ctx, entityID)
 }
